Split PanelSections component out.
